extract dbhelper from sync-pending-events
